pyproject.toml
[project]
name = "mba-ia-intro"
version = "0.1.0"
description = "Material de estudo para o MBA de Inteligência Artificial"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "langchain>=1.2.8",
    "langchain-classic>=1.0.1",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.0",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.7",
    "langchain-postgres>=0.0.16",
    "langchain-text-splitters>=1.1.0",
    "pypdf>=6.6.2",
    "python-dotenv>=1.2.1",
]

[tool.ruff]
line-length = 120
